+++ src/components/AppNavigation.vue
@@ -1,5 +1,9 @@
 <template>
-  <NcAppNavigation>
+  <NcAppNavigation
+    :class="{
+      '[&.app-navigation--close]:transform-none [&:not(.app-navigation--close)]:ml-[unset]': addNC25ShimClass
+    }"
+  >
     <template #list>
       <AccountTypeListItem
         v-for="accountType in accountTypes"
@@ -42,7 +46,7 @@
 
 <script setup lang="ts">
 
-  import { computed } from 'vue';
+  import { computed, onMounted, ref } from 'vue';
 
   import NcAppNavigation from '@nextcloud/vue/dist/Components/NcAppNavigation';
   import NcAppNavigationItem from '@nextcloud/vue/dist/Components/NcAppNavigationItem';
@@ -75,5 +79,12 @@
 
   // TODO: initialize accounts at some other point
   accountService.fetchAccounts();
+
+  const addNC25ShimClass = ref(false);
+  onMounted(() => {
+    if (window.getComputedStyle(document.body).position === 'fixed') {
+      addNC25ShimClass.value = true;
+    }
+  });
 
 </script>