+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+### Changed
+
+- Fixed navigation bar appearance in NC 25
+
 ## [0.11.0] - 2022-11-30
 
 ### Added